Удалён мусор; к стейтам больше напрямую не обращаемся, только через геттеры; использован bem-модификатор у &--was-hit и &--was-missed; модули actions, getters, mutations, state теперь в одном файле index.
