--- vue/src/store/getters.js
@@ -1,3 +1,0 @@
-export default {
-    isGameActive: state => state.gameStarted && !state.gameOver,
-}
