--- vue/src/store/state.js
@@ -1,18 +1,0 @@
-export default {
-    gameStarted: false,
-    gameOver: false,
-    score: 0,
-    comboCount: 0,
-    comboBonus: 0,
-    misses: 0,
-    maxMisses: 5,
-    statusMessage: '',
-    activeCell: null,
-    moleWasHit: false,
-    moleWasMissed: false,
-    moleTimeout: null,
-    moleInterval: null,
-    statusTimeout: null,
-    missedIndex: null,
-    hitIndex: null,
-}
